Remove functions git.repos and git.status
Reason: these functions require knowledge of the
root of the bryan.* modules and this repo should
be ignorant of this.

It now knows the location of this root via the
environment variable [$__BRYAN_ROOT__] (which
is itself set by [~/.profile], but we do not
want to be coupled to that variable name.

Do this because at time of writing:

    (i) this repo is named according to the 'bryan submodule'
    name scheme 'bryan.NNN.<name>' and installed locally by
    {~/.profile} under [~/.BRYAN]i, but

    (ii) we intent to rename this (and other such) repos
    to 'bch.NNN.<name>' and have them installed locally
    under [~/.BCH].